remove unused imports and move parsing tests to config.rs
